+++ client/coral-framework/reducers/index.js
@@ -3,7 +3,7 @@ import user from './user';
 import asset from './asset';
 import {reducer as commentBox} from '../../coral-plugin-commentbox';
 import {pluginReducers} from '../helpers/plugins';
-console.log(pluginReducers);
+
 export default {
   auth,
   user,
