+++ plugins/coral-plugin-offtopic/client/components/OffTopicFilter.js
@@ -1,12 +1,7 @@
 import React from 'react';
 import styles from './styles.css';
-import {connect} from 'react-redux';
-import {bindActionCreators} from 'redux';
-import {toggleCheckbox} from '../actions';
-import {addClassName, removeClassName} from 'coral-embed-stream/src/actions/comment';
-import {closeViewingOptions} from 'coral-embed-stream/src/actions/stream';
 
-class OffTopicFilter extends React.Component {
+export default class OffTopicFilter extends React.Component {
 
   tag = 'OFF_TOPIC';
   className = 'offTopicComment';
@@ -35,13 +30,3 @@ class OffTopicFilter extends React.Component {
     );
   }
 }
-
-const mapStateToProps = ({comment, offTopic}) => ({
-  classNames: comment.classNames,
-  checked: offTopic.checked
-});
-
-const mapDispatchToProps = (dispatch) =>
-  bindActionCreators({addClassName, removeClassName, toggleCheckbox, closeViewingOptions}, dispatch);
-
-export default connect(mapStateToProps, mapDispatchToProps)(OffTopicFilter);
