+++ plugins/coral-plugin-offtopic/client/constants.js
@@ -1,1 +1,1 @@
-export const TOGGLE_CHECKBOX = "TOGGLE_CHECKBOX";
+export const TOGGLE_CHECKBOX = 'TOGGLE_CHECKBOX';
